pyproject.toml
[project]
name = "matching-arts-v3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clip",
    "fastapi>=0.115.12",
    "ftfy>=6.3.1",
    "joblib>=1.4.2",
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "polars>=1.26.0",
    "pydantic>=2.11.2",
    "regex>=2024.11.6",
    "scikit-learn>=1.6.1",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.34.0",
]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }
